Huge redesign.
All requests go through a JSON API. Everything else is static HTML or
Javascript.

After this, I don’t think this is the right way to go, since JSON
requests are asynchronous, and things like the user should just be put
into the templates on the server side rather than fetching that
information on the client. It’s a cleaner design, but not very
efficient.

Also, fetch the angular code remotely, rather than from a local copy.